media_analyzer/pyproject.toml
[build-system]
requires = ["setuptools >= 76.1.0"]
build-backend = "setuptools.build_meta"

[project]
name = "media_analyzer"
version = "1.0.0"

description = """
Analyzing video reels to get recipes
"""

keywords = ["Machinelearning","Media","Analysis"]

dependencies = [
    "google-genai==1.7.0",
]

requires-python = ">=3.9"

readme = {file = "README.md", content-type = "text/markdown"}
authors = [
    {name = "Sidhant Raj Khati",email = "sidhant.raj.khati@gmail.com"}
]

license = { text = "MIT" }

classifiers = [
   "Programming Language :: Python :: 3",
   "Operating System :: OS Independent",
]

[tool.setuptools.packages]
find = {}